+++ libs/gi/assets/src/gen/weapons/SurfsUp/UI_EquipIcon_Catalyst_MechaPufferfish_Awaken.png
