+++ libs/gi/assets/src/gen/weapons/SurfsUp/index.ts
