+++ libs/gi/dm-localization/assets/locales/chs/weaponNames_gen.json
@@ -151,7 +151,7 @@
   "CashflowSupervision": "金流监督",
   "TomeOfTheEternalFlow": "万世流涌大典",
   "CranesEchoingCall": "鹤鸣余音",
-  "SurfingTime": "冲浪时光",
+  "SurfsUp": "冲浪时光",
   "HuntersBow": "猎弓",
   "SeasonedHuntersBow": "历练的猎弓",
   "RavenBow": "鸦羽弓",
