+++ libs/gi/dm-localization/assets/locales/cht/weaponNames_gen.json
@@ -151,7 +151,7 @@
   "CashflowSupervision": "金流監督",
   "TomeOfTheEternalFlow": "萬世流湧大典",
   "CranesEchoingCall": "鶴鳴餘音",
-  "SurfingTime": "衝浪時光",
+  "SurfsUp": "衝浪時光",
   "HuntersBow": "獵弓",
   "SeasonedHuntersBow": "歷練的獵弓",
   "RavenBow": "鴉羽弓",
