+++ libs/gi/dm-localization/assets/locales/de/weaponNames_gen.json
@@ -151,7 +151,7 @@
   "CashflowSupervision": "Monetäre Aufsichtseinheit",
   "TomeOfTheEternalFlow": "Lehre des ewigen Flusses",
   "CranesEchoingCall": "Widerhallender Ruf des Kranichs",
-  "SurfingTime": "Zeit fürs Wellenreiten",
+  "SurfsUp": "Zeit fürs Wellenreiten",
   "HuntersBow": "Jagdbogen",
   "SeasonedHuntersBow": "Alter Jagdbogen",
   "RavenBow": "Rabenbogen",
