+++ libs/gi/dm-localization/assets/locales/es/weaponNames_gen.json
@@ -151,7 +151,7 @@
   "CashflowSupervision": "Supervisor Flujoáurico",
   "TomeOfTheEternalFlow": "Escrituras del Fluir Sempiterno",
   "CranesEchoingCall": "Reverberación de la Grulla",
-  "SurfingTime": "Hora de Surfear",
+  "SurfsUp": "Hora de Surfear",
   "HuntersBow": "Arco del Cazador",
   "SeasonedHuntersBow": "Arco del Cazador Estacional",
   "RavenBow": "Arco de Cuervo",
