+++ libs/gi/dm-localization/assets/locales/fr/weaponNames_gen.json
@@ -151,7 +151,7 @@
   "CashflowSupervision": "Supervision de trésorerie",
   "TomeOfTheEternalFlow": "Tome du flux éternel",
   "CranesEchoingCall": "Écho de la grue",
-  "SurfingTime": "Instant surfant",
+  "SurfsUp": "Instant surfant",
   "HuntersBow": "Arc de chasse",
   "SeasonedHuntersBow": "Arc de chasse aguerri",
   "RavenBow": "Arc du corbeau",
