+++ libs/gi/dm-localization/assets/locales/fr/weapon_SurfsUp_gen.json
