+++ libs/gi/dm-localization/assets/locales/it/weaponNames_gen.json
@@ -151,7 +151,7 @@
   "CashflowSupervision": "Supervisione pecuniaria",
   "TomeOfTheEternalFlow": "Tomo del flusso perenne",
   "CranesEchoingCall": "Canto della gru echeggiante",
-  "SurfingTime": "Vai col surf",
+  "SurfsUp": "Vai col surf",
   "HuntersBow": "Arco da caccia",
   "SeasonedHuntersBow": "Arco da caccia per esperti",
   "RavenBow": "Arco corvino",
