+++ libs/gi/dm-localization/assets/locales/ja/weaponNames_gen.json
@@ -151,7 +151,7 @@
   "CashflowSupervision": "凛流の監視者",
   "TomeOfTheEternalFlow": "久遠流転の大典",
   "CranesEchoingCall": "鶴鳴の余韻",
-  "SurfingTime": "サーフィンタイム",
+  "SurfsUp": "サーフィンタイム",
   "HuntersBow": "狩猟弓",
   "SeasonedHuntersBow": "歴戦の狩猟弓",
   "RavenBow": "鴉羽の弓",
