+++ libs/gi/dm-localization/assets/locales/ko/weaponNames_gen.json
@@ -151,7 +151,7 @@
   "CashflowSupervision": "현금 흐름 감독",
   "TomeOfTheEternalFlow": "영원히 샘솟는 법전",
   "CranesEchoingCall": "학의 여음",
-  "SurfingTime": "서핑 타임",
+  "SurfsUp": "서핑 타임",
   "HuntersBow": "사냥활",
   "SeasonedHuntersBow": "노련의 사냥활",
   "RavenBow": "까마귀깃 활",
