+++ libs/gi/dm-localization/assets/locales/pt/weaponNames_gen.json
@@ -151,7 +151,7 @@
   "CashflowSupervision": "Supervisão de Caixa",
   "TomeOfTheEternalFlow": "Rito do Fluxo Eterno",
   "CranesEchoingCall": "Chamado Ecoante da Garça",
-  "SurfingTime": "Hora do Surfe",
+  "SurfsUp": "Hora do Surfe",
   "HuntersBow": "Arco do Caçador",
   "SeasonedHuntersBow": "Arco do Caçador Experiente",
   "RavenBow": "Arco de Corvo",
