+++ libs/gi/dm-localization/assets/locales/ru/weaponNames_gen.json
@@ -151,7 +151,7 @@
   "CashflowSupervision": "Казначейский надзор",
   "TomeOfTheEternalFlow": "Обряд вечного течения",
   "CranesEchoingCall": "Звонкий клич журавля",
-  "SurfingTime": "Лови волну",
+  "SurfsUp": "Лови волну",
   "HuntersBow": "Лук охотника",
   "SeasonedHuntersBow": "Лук опытного охотника",
   "RavenBow": "Лук ворона",
