+++ libs/gi/dm-localization/assets/locales/th/weaponNames_gen.json
@@ -151,7 +151,7 @@
   "CashflowSupervision": "Cashflow Supervision",
   "TomeOfTheEternalFlow": "Tome of the Eternal Flow",
   "CranesEchoingCall": "Crane's Echoing Call",
-  "SurfingTime": "Surf's Up",
+  "SurfsUp": "Surf's Up",
   "HuntersBow": "Hunter's Bow",
   "SeasonedHuntersBow": "Seasoned Hunter's Bow",
   "RavenBow": "Raven Bow",
