+++ libs/gi/dm-localization/assets/locales/tr/weaponNames_gen.json
@@ -151,7 +151,7 @@
   "CashflowSupervision": "Nakit Akışı İzleyici",
   "TomeOfTheEternalFlow": "Sonsuz Akış Kitabı",
   "CranesEchoingCall": "Turnam Selam Götür",
-  "SurfingTime": "Sörf Zamanı",
+  "SurfsUp": "Sörf Zamanı",
   "HuntersBow": "Avcı Yayı",
   "SeasonedHuntersBow": "Usta Avcı Yayı",
   "RavenBow": "Kuzgun Yayı",
