+++ libs/gi/dm-localization/assets/locales/vi/weaponNames_gen.json
@@ -151,7 +151,7 @@
   "CashflowSupervision": "Quản Đốc Vàng Ròng",
   "TomeOfTheEternalFlow": "Nghi Thức Dòng Chảy Vĩnh Hằng",
   "CranesEchoingCall": "Dư Âm Tiếng Hạc",
-  "SurfingTime": "Thời Khắc Lướt Sóng",
+  "SurfsUp": "Thời Khắc Lướt Sóng",
   "HuntersBow": "Cung Săn Bắn",
   "SeasonedHuntersBow": "Cung Săn Bắn Tôi Luyện",
   "RavenBow": "Cung Lông Quạ",
