+++ libs/gi/sheets/src/Weapons/Catalyst/SurfsUp/index.tsx
@@ -14,7 +14,7 @@ import type { IWeaponSheet } from '../../IWeaponSheet'
 import { WeaponSheet, headerTemplate } from '../../WeaponSheet'
 import { dataObjForWeaponSheet } from '../../util'
 
-const key: WeaponKey = 'SurfingTime'
+const key: WeaponKey = 'SurfsUp'
 const data_gen = allStats.weapon.data[key]
 const [, trm] = trans('weapon', key)
