+++ libs/gi/stats/src/allStat_gen.json
@@ -55204,7 +55204,7 @@
           ]
         }
       },
-      "SurfingTime": {
+      "SurfsUp": {
         "weaponType": "catalyst",
         "rarity": 5,
         "mainStat": {
